Relationship Completed

+++ src/tasks/tasks.service.ts
@@ -19,8 +19,9 @@ export class TasksService {
     return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user:User): Promise<Task> {
+    const found = await this.taskRepository.findOne({where: {id,userId: user.id}});
+    console.log(found)
 
     if (!found) {
       throw new NotFoundException('Task with id Not Found!');
@@ -36,16 +37,16 @@ export class TasksService {
     return this.taskRepository.createTask(title, description, user);
   }
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user:User): Promise<void> {
+    const result = await this.taskRepository.delete({id, userId:user.id});
 
     if (!result.affected) {
       throw new NotFoundException('Task with id Not Found!');
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user:User): Promise<Task> {
+    const task = await this.getTaskById(id,user);
     task.status = status;
     await task.save();
     return task;
